# backend/app/Http/Controllers/Api/V1/EleveController.php
<?php
namespace App\Http\Controllers\Api\V1;

use App\Exports\ElevesExport;
use App\Http\Controllers\Api\BaseApiController;
use App\Http\Requests\Eleve\StoreEleveRequest;
use App\Http\Requests\Eleve\UpdateEleveRequest;
use App\Http\Resources\EleveDetailResource;
use App\Models\{Eleve, ParentEleve};
use App\Services\{EleveService};
use Illuminate\Http\{Request, JsonResponse};
use Illuminate\Support\Facades\{DB, Storage};
use Maatwebsite\Excel\Facades\Excel;

class EleveController extends BaseApiController
{
    protected string $model      = Eleve::class;
    protected array  $with       = ['wilaya', 'parents'];
    protected array  $searchable = ['nom', 'prenom', 'numero_inscription'];
    protected array  $filterable = ['statut', 'niveau_scolaire', 'sexe', 'wilaya_id'];
    protected array  $sortable   = ['nom', 'prenom', 'created_at', 'niveau_scolaire'];
    protected int    $perPage    = 20;

    /** Un parent/enseignant ne liste que les élèves de son périmètre. */
    protected ?string $colonnePerimetreEleve = 'id';

    public function __construct(
        private EleveService $eleveService
    ) {}

    /**
     * @OA\Get(
     *     path="/api/v1/eleves",
     *     summary="Liste paginée des élèves",
     *     tags={"Eleves"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(ref="#/components/parameters/TenantId"),
     *     @OA\Parameter(name="per_page",       in="query", required=false, @OA\Schema(type="integer", default=20)),
     *     @OA\Parameter(name="statut",         in="query", required=false, @OA\Schema(type="string",  enum={"actif","inactif","suspendu"})),
     *     @OA\Parameter(name="niveau_scolaire",in="query", required=false, @OA\Schema(type="string")),
     *     @OA\Parameter(name="search",         in="query", required=false, @OA\Schema(type="string")),
     *     @OA\Response(
     *         response=200,
     *         description="Liste des élèves",
     *         @OA\JsonContent(
     *             @OA\Property(property="success", type="boolean"),
     *             @OA\Property(property="data", type="object",
     *                 @OA\Property(property="data",  type="array", @OA\Items(ref="#/components/schemas/Eleve")),
     *                 @OA\Property(property="meta",  ref="#/components/schemas/PaginationMeta"),
     *                 @OA\Property(property="stats", type="object")
     *             )
     *         )
     *     )
     * )
     */
    public function index(Request $request): JsonResponse
    {
        $perPage = $request->per_page ?? $this->perPage;

        $query = $this->indexQuery($request)
            ->with([
                'wilaya:id,nom_fr',
                'commune:id,nom_fr',
                'parents:id,nom,prenom,telephone_1,email',
                'diagnosticEleve:id,eleve_id,score_risque,niveau_global',
                'inscriptions' => fn($q) => $q->where('statut', 'validée')
                    ->with('groupe:id,nom,matiere_id')
                    ->select('id', 'eleve_id', 'groupe_id', 'statut'),
            ])
            ->withCount([
                'inscriptions as nb_inscriptions' => fn($q) => $q->where('statut', 'validée'),
                'presences as nb_presences',
                'presences as nb_absences' => fn($q) => $q->where('statut', 'absent'),
            ]);

        if (!in_array($request->sort ?? 'created_at', $this->sortable)) {
            $query->orderBy('created_at', 'desc');
        }

        $cursor = $query->cursorPaginate($perPage);
        $total  = $query->toBase()->count();

        // Les statistiques globales n'ont de sens que pour les rôles ayant
        // une vision complète du tenant ; sinon elles fuiteraient des
        // effectifs hors périmètre.
        $perimetre = app(\App\Services\PerimetreAccesService::class);

        $stats = !$perimetre->aVisionGlobale(auth('api')->user()) ? null : cache()->remember(
            "eleves_stats_" . config('tenant.current_id'),
            300,
            fn() => [
                'total'    => Eleve::count(),
                'actifs'   => Eleve::where('statut', 'actif')->count(),
                'nouveaux' => Eleve::whereMonth('created_at', now()->month)->count(),
            ]
        );

        $itemCount = $cursor->count();

        return response()->json([
            'success' => true,
            'message' => 'Élèves récupérés',
            'data'    => $cursor->items(),
            'meta'    => [
                'total'        => $total,
                'per_page'     => (int) $perPage,
                'current_page' => 1,
                'last_page'    => (int) ceil($total / $perPage),
                'has_more'     => $cursor->hasMorePages(),
                'stats'        => $stats,
            ],
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/v1/eleves",
     *     summary="Créer un élève",
     *     tags={"Eleves"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(ref="#/components/parameters/TenantId"),
     *     @OA\RequestBody(required=true, @OA\JsonContent(ref="#/components/schemas/EleveInput")),
     *     @OA\Response(response=201, description="Élève créé",      @OA\JsonContent(ref="#/components/schemas/SuccessResponse")),
     *     @OA\Response(response=422, description="Données invalides",@OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function store(StoreEleveRequest $request): JsonResponse
    {
        $eleve = DB::transaction(function () use ($request) {
            $numero = $this->eleveService->genererNumero();

            $eleve = Eleve::create([
                ...$request->validated(),
                'numero_inscription' => $numero,
                'tenant_id'          => config('tenant.current_id'),
            ]);

            if ($request->has('parents')) {
                foreach ($request->parents as $index => $parentData) {
                    $parent = ParentEleve::firstOrCreate(
                        ['telephone_1' => $parentData['telephone_1'], 'tenant_id' => config('tenant.current_id')],
                        [...$parentData, 'tenant_id' => config('tenant.current_id')]
                    );
                    $eleve->parents()->attach($parent->id, ['est_principal' => $index === 0]);
                }
            }

            $this->eleveService->genererQRCode($eleve);
            return $eleve;
        });

        cache()->forget("eleves_stats_" . config('tenant.current_id'));

        return $this->created(
            $eleve->load(['wilaya', 'commune', 'parents']),
            "Élève {$eleve->nom} {$eleve->prenom} inscrit avec succès"
        );
    }

    /**
     * @OA\Get(
     *     path="/api/v1/eleves/{id}",
     *     summary="Détail d'un élève",
     *     tags={"Eleves"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(ref="#/components/parameters/TenantId"),
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="string", format="uuid")),
     *     @OA\Response(response=200, description="Détail élève",   @OA\JsonContent(ref="#/components/schemas/SuccessResponse")),
     *     @OA\Response(response=404, description="Non trouvé",     @OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function show(string $id): JsonResponse
    {
        $eleve = Eleve::with([
            'wilaya:id,nom_fr',
            'commune:id,nom_fr',
            'inscriptions' => fn($q) => $q->where('statut', 'validée')
                ->with('groupe:id,nom,matiere_id'),
            'parents:id,nom,prenom,telephone_1,email,lien',
            'notes' => fn($q) => $q->whereNotNull('note')
                ->latest()->limit(50)
                ->with('evaluation:id,type_eval,trimestre,note_sur,coefficient,date_evaluation'),
            'absencesJournalieres' => fn($q) => $q->latest()->limit(20),
            'diagnosticEleve:id,eleve_id,score_risque,niveau_global,matieres_en_danger',
            'factures' => fn($q) => $q->whereIn('statut', ['émise', 'en_retard', 'partiellement_payée'])
                ->with('paiements'),
        ])
        ->withCount([
            'presences',
            'presences as presences_presentes' => fn($q) => $q->whereIn('statut', ['présent', 'retard']),
            'factures as factures_impayees'    => fn($q) => $q->whereNotIn('statut', ['payée', 'annulée']),
        ])
        ->findOrFail($id);

        // Le scope tenant garantit déjà l'isolation inter-établissements ;
        // ici on vérifie le périmètre INTRA-tenant (parent -> ses enfants,
        // enseignant -> ses groupes).
        if ($reponse = $this->verifierPerimetreEleve($eleve->id)) {
            return $reponse;
        }

        $stats = $this->eleveService->getStatsAcademiques($eleve);

        return $this->success([
            'eleve'       => new EleveDetailResource($eleve),
            'statistiques'=> $stats,
        ]);
    }

    /**
     * @OA\Put(
     *     path="/api/v1/eleves/{id}",
     *     summary="Mettre à jour un élève",
     *     tags={"Eleves"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(ref="#/components/parameters/TenantId"),
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="string", format="uuid")),
     *     @OA\RequestBody(required=true, @OA\JsonContent(ref="#/components/schemas/EleveInput")),
     *     @OA\Response(response=200, description="Élève mis à jour", @OA\JsonContent(ref="#/components/schemas/SuccessResponse")),
     *     @OA\Response(response=404, description="Non trouvé",       @OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function update(UpdateEleveRequest $request, string $id): JsonResponse
    {
        $eleve = Eleve::findOrFail($id);
        $eleve->update($request->validated());

        cache()->forget("eleves_stats_" . config('tenant.current_id'));

        return $this->success(
            $eleve->fresh(['wilaya', 'commune', 'parents']),
            'Élève mis à jour avec succès'
        );
    }

    /**
     * @OA\Delete(
     *     path="/api/v1/eleves/{id}",
     *     summary="Supprimer un élève (soft delete)",
     *     tags={"Eleves"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(ref="#/components/parameters/TenantId"),
     *     @OA\Parameter(name="id", in="path", required=true, @OA\Schema(type="string", format="uuid")),
     *     @OA\Response(response=200, description="Supprimé",    @OA\JsonContent(ref="#/components/schemas/SuccessResponse")),
     *     @OA\Response(response=404, description="Non trouvé",  @OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function destroy(string $id): JsonResponse
    {
        $eleve = Eleve::findOrFail($id);

        $impayes = $eleve->factures()
            ->whereNotIn('statut', ['payée', 'annulée'])
            ->count();

        if ($impayes > 0) {
            return $this->error(
                "Impossible de supprimer : {$impayes} facture(s) impayée(s)",
                'HAS_UNPAID_INVOICES', 422
            );
        }

        $nom = "{$eleve->nom} {$eleve->prenom}";
        $eleve->update(['statut' => 'inactif']);
        $eleve->delete();

        return $this->success(null, "{$nom} a été archivé");
    }

    public function uploadPhoto(Request $request, string $id): JsonResponse
    {
        $request->validate(['photo' => 'required|image|mimes:jpg,jpeg,png,webp|max:2048']);

        $eleve = Eleve::findOrFail($id);
        if ($eleve->photo_url) {
            Storage::disk('public')->delete($eleve->photo_url);
        }

        $path = $request->file('photo')->store("photos/eleves/{$eleve->tenant_id}", 'public');
        $eleve->update(['photo_url' => $path]);

        return $this->success(['photo_url' => Storage::url($path)], 'Photo mise à jour');
    }

    public function notes(Request $request, string $id): JsonResponse
    {
        $eleve = Eleve::findOrFail($id);

        if ($reponse = $this->verifierPerimetreEleve($eleve->id)) {
            return $reponse;
        }

        $notes = $eleve->notes()
            ->with(['evaluation' => fn($q) => $q->with('groupe.matiere:id,nom_fr,couleur,coefficient')])
            ->when($request->trimestre, fn($q) => $q->whereHas('evaluation', fn($eq) => $eq->where('trimestre', $request->trimestre)))
            ->when($request->groupe_id, fn($q) => $q->whereHas('evaluation', fn($eq) => $eq->where('groupe_id', $request->groupe_id)))
            ->get();

        $parMatiere = $notes->groupBy(fn($n) => $n->evaluation->groupe->matiere->nom_fr)
            ->map(fn($groupNotes, $matiere) => [
                'matiere'     => $matiere,
                'couleur'     => $groupNotes->first()->evaluation->groupe->matiere->couleur ?? '#1E5EBC',
                'coefficient' => $groupNotes->first()->evaluation->groupe->matiere->coefficient,
                'notes'       => $groupNotes->map(fn($n) => [
                    'id'       => $n->id,
                    'note'     => $n->note,
                    'note_sur' => $n->evaluation->note_sur,
                    'appreciation' => $n->appreciation,
                    'type'     => $n->evaluation->type_eval,
                    'date'     => $n->evaluation->date_evaluation,
                    'absent'   => $n->absent,
                ])->values(),
                'moyenne' => $groupNotes->whereNotNull('note')->avg(fn($n) => ($n->note / $n->evaluation->note_sur) * 20),
            ])->values();

        return $this->success([
            'notes'            => $parMatiere,
            'moyenne_generale' => $this->eleveService->calculerMoyenne($eleve->id, $request->groupe_id, $request->trimestre),
            'taux_presence'    => $this->eleveService->calculerTauxPresence($eleve->id),
        ]);
    }

    public function presences(Request $request, string $id): JsonResponse
    {
        $eleve = Eleve::findOrFail($id);

        if ($reponse = $this->verifierPerimetreEleve($eleve->id)) {
            return $reponse;
        }

        $presences = $eleve->presences()
            ->with(['seance' => fn($q) => $q->with([
                'cours.groupe.matiere:id,nom_fr,couleur',
                'cours.enseignant:id,nom,prenom',
            ])])
            ->when($request->mois, fn($q) => $q->whereMonth('created_at', $request->mois))
            ->when($request->annee, fn($q) => $q->whereYear('created_at', $request->annee))
            ->orderByDesc('created_at')
            ->paginate(20);

        $statsPresence = [
            'total'   => $eleve->presences()->count(),
            'presents'=> $eleve->presences()->whereIn('statut', ['présent','retard'])->count(),
            'absents' => $eleve->presences()->where('statut', 'absent')->count(),
            'taux'    => $this->eleveService->calculerTauxPresence($eleve->id),
        ];

        return $this->paginatedResponse($presences, 'Présences récupérées', ['stats' => $statsPresence]);
    }

    public function paiements(string $id): JsonResponse
    {
        $eleve = Eleve::findOrFail($id);

        if ($reponse = $this->verifierPerimetreEleve($eleve->id)) {
            return $reponse;
        }

        $totalPaye  = $eleve->paiements()
            ->where('statut', 'confirmé')->sum('montant');
        $totalDette = max(0, $eleve->factures()
            ->whereNotIn('statut', ['payée', 'annulée'])->sum('total_ttc')
            - $eleve->paiements()
                ->whereHas('facture', fn($q) => $q->whereNotIn('statut', ['payée', 'annulée']))
                ->where('statut', 'confirmé')
                ->sum('montant'));

        return $this->success([
            'factures'  => $eleve->factures()->with('paiements', 'lignes')->orderByDesc('date_emission')->get(),
            'financier' => [
                'total_paye'  => $totalPaye,
                'total_dette' => $totalDette,
                'nb_factures' => $eleve->factures()->count(),
                'nb_impayes'  => $eleve->factures()->whereNotIn('statut', ['payée', 'annulée'])->count(),
            ],
        ]);
    }

    public function bulletins(string $id): JsonResponse
    {
        $eleve = Eleve::findOrFail($id);

        if ($reponse = $this->verifierPerimetreEleve($eleve->id)) {
            return $reponse;
        }
        return $this->success(
            $eleve->bulletins()->with('groupe')->orderByDesc('created_at')->get()
        );
    }

    public function statistiques(string $id): JsonResponse
    {
        $eleve = Eleve::withCount([
            'inscriptions',
            'presences as total_presences' => fn($q) => $q->whereIn('statut', ['présent', 'retard']),
        ])->findOrFail($id);

        if ($reponse = $this->verifierPerimetreEleve($eleve->id)) {
            return $reponse;
        }

        return $this->success($eleve);
    }

    public function inscrire(Request $request, string $id): JsonResponse
    {
        $validated = $request->validate([
            'groupe_id'      => 'required|uuid|exists:groupes,id',
            'annee_scolaire' => 'nullable|string|regex:/^\d{4}-\d{4}$/',
            'date_inscription'=> 'nullable|date',
        ]);

        $eleve  = Eleve::findOrFail($id);
        $groupe = \App\Models\Groupe::findOrFail($validated['groupe_id']);

        $dejaInscrit = $eleve->inscriptions()
            ->where('groupe_id', $groupe->id)
            ->where('statut', 'validée')
            ->exists();

        if ($dejaInscrit) {
            return $this->error("L'élève est déjà inscrit dans ce groupe", 'ALREADY_ENROLLED', 409);
        }

        $nbInscrits = $groupe->inscriptions()->where('statut', 'validée')->count();
        if ($nbInscrits >= $groupe->capacite_max) {
            return $this->error("Le groupe est complet ({$groupe->capacite_max} max)", 'GROUP_FULL', 409);
        }

        $inscription = $eleve->inscriptions()->create([
            'tenant_id'       => config('tenant.current_id'),
            'groupe_id'       => $groupe->id,
            'date_inscription'=> $validated['date_inscription'] ?? now()->toDateString(),
            'statut'          => 'validée',
            'inscrit_par'     => auth('api')->id(),
        ]);

        return $this->created($inscription->load('groupe'), "Inscription au groupe {$groupe->nom} validée");
    }

    public function import(Request $request): JsonResponse
    {
        $request->validate(['fichier' => 'required|file|mimes:xlsx,xls,csv|max:10240']);

        // Stub — à implémenter avec maatwebsite/excel
        $request->file('fichier')->store('imports');
        return $this->success(null, 'Import en cours de traitement');
    }

    public function export(Request $request)
    {
        $eleves = Eleve::with(['wilaya', 'parents'])->get();
        return $this->success($eleves);
    }

    public function exportExcel(): \Symfony\Component\HttpFoundation\BinaryFileResponse
    {
        return Excel::download(
            new ElevesExport(config('tenant.current_id')),
            'eleves.xlsx'
        );
    }
}