# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduGest DZ API",
        "description": "API complète — plateforme SaaS de gestion scolaire pour l'Algérie.\nAuthentification via JWT Bearer (header Authorization: Bearer <token>).\nMulti-tenant : chaque requête doit inclure le header X-Tenant-ID.",
        "contact": {
            "email": "support@edugest.dz"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur courant"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/eleves": {
            "get": {
                "tags": [
                    "Eleves"
                ],
                "summary": "Liste paginée des élèves",
                "operationId": "3af868b3bb89258153c28748f2a1927c",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantId"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "inactif",
                                "suspendu"
                            ]
                        }
                    },
                    {
                        "name": "niveau_scolaire",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des élèves",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Eleve"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                },
                                                "stats": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Eleves"
                ],
                "summary": "Créer un élève",
                "operationId": "7aa5189f2897dfec97cfdff09526e306",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EleveInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Élève créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/eleves/{id}": {
            "get": {
                "tags": [
                    "Eleves"
                ],
                "summary": "Détail d'un élève",
                "operationId": "2259e49f74aafd71da41b53ede3e2875",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantId"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail élève",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Eleves"
                ],
                "summary": "Mettre à jour un élève",
                "operationId": "071fa64224794ff23bc831bb32641e5a",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantId"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EleveInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Élève mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Eleves"
                ],
                "summary": "Supprimer un élève (soft delete)",
                "operationId": "dea9a69cdf3567d07caceac09fafd30f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantId"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 97
                    }
                },
                "type": "object"
            },
            "Eleve": {
                "required": [
                    "id",
                    "nom",
                    "prenom",
                    "niveau_scolaire",
                    "statut"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Benali"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Amira"
                    },
                    "date_naissance": {
                        "type": "string",
                        "format": "date",
                        "example": "2010-03-15"
                    },
                    "sexe": {
                        "type": "string",
                        "enum": [
                            "M",
                            "F"
                        ]
                    },
                    "niveau_scolaire": {
                        "type": "string",
                        "example": "3AS"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "inactif",
                            "suspendu"
                        ]
                    },
                    "photo_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "tenant_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EleveInput": {
                "required": [
                    "nom",
                    "prenom",
                    "niveau_scolaire",
                    "date_naissance",
                    "sexe"
                ],
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "Benali"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Amira"
                    },
                    "date_naissance": {
                        "type": "string",
                        "format": "date",
                        "example": "2010-03-15"
                    },
                    "sexe": {
                        "type": "string",
                        "enum": [
                            "M",
                            "F"
                        ]
                    },
                    "niveau_scolaire": {
                        "type": "string",
                        "example": "3AS"
                    },
                    "wilaya_id": {
                        "type": "integer",
                        "example": 31
                    },
                    "commune_id": {
                        "type": "integer",
                        "example": 310
                    },
                    "adresse": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "TenantId": {
                "name": "X-Tenant-ID",
                "in": "header",
                "description": "Identifiant unique du tenant (école/centre)",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}